Remove updateAllReachDistances() from API, AWApiReachHelper and move to ReachUpdater
This method doesn't do any network traffic, it only updates local data based on the user's location.

It should still go away completely -- we should seek to avoid pre-baking distance.
