+++ American Whitewater/ViewControllers/Favorites/FavoritesViewController.swift
@@ -9,8 +9,7 @@ class FavoritesViewController: UIViewController {
     
     private let managedObjectContext = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
     private var fetchedResultsController: NSFetchedResultsController<Reach>?
-    
-    private let reachUpdater = ReachUpdater()
+    private lazy var reachUpdater = ReachUpdater(managedObjectContext: managedObjectContext)
     
     override func viewDidLoad() {
         super.viewDidLoad()
