+++ American Whitewater/ViewControllers/Reaches/RunsListViewController.swift
@@ -6,9 +6,9 @@ import CoreLocation
 import KeychainSwift
 
 class RunsListViewController: UIViewController {
-    private let reachUpdater = ReachUpdater()
     private let managedObjectContext = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
     private var fetchedResultsController: NSFetchedResultsController<Reach>?
+    private lazy var reachUpdater = ReachUpdater(managedObjectContext: managedObjectContext)
     
     @IBOutlet weak var tableView: UITableView!
     @IBOutlet weak var runnableFilterContainerView: UIView!
